PR2: Add support for manipulating extension fields.
MapExtFieldPtr takes a parameter name and returns an opaque field
token. In agreement with FTE this token contains some random bits
that are verified and stripped on use to avoid mods hardcoding field
offsets, which would over time lead to engine incompatibilities.

SetExtFieldPtr takes an entnum, a field token, a pointer to some
data as well as size which, if token and size is valid, performs a
memcpy of the data to the field referenced.

GetExtFieldPtr is the inverse of SetExtFieldPtr with same parameters.

Availability of above functions is determined via Map_Extension.